pyproject.toml
[project]
name = "matriculaciones"
version = "0.1.0"
description = "Script CLI para automatizar matriculaciones de estudiantes en Moodle desde PocketBase"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
]

[project.scripts]
matriculaciones = "matriculaciones.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matriculaciones"]
